requirements.txt
git+https://github.com/django/django.git@stable/1.7.x#egg=django
django-braces
django-debug-toolbar
git+https://github.com/alex/django-filter.git@develop#egg=django_filter